PIM-7464: Fix memory leak during rule execution

+++ src/PimEnterprise/Component/CatalogRule/Engine/ProductRuleApplier.php
@@ -16,6 +16,7 @@ use Akeneo\Bundle\RuleEngineBundle\Event\RuleEvents;
 use Akeneo\Bundle\RuleEngineBundle\Event\SelectedRuleEvent;
 use Akeneo\Bundle\RuleEngineBundle\Model\RuleInterface;
 use Akeneo\Bundle\RuleEngineBundle\Model\RuleSubjectSetInterface;
+use Akeneo\Component\StorageUtils\Cache\CacheClearerInterface;
 use Akeneo\Component\StorageUtils\Cursor\PaginatorFactoryInterface;
 use Akeneo\Component\StorageUtils\Detacher\ObjectDetacherInterface;
 use PimEnterprise\Component\CatalogRule\Engine\ProductRuleApplier\ProductsSaver;
@@ -48,13 +49,25 @@ class ProductRuleApplier implements ApplierInterface
     /** @var ObjectDetacherInterface */
     protected $objectDetacher;
 
+    /** @var int */
+    protected $pageSize;
+
+    /** @var CacheClearerInterface|null */
+    protected $cacheClearer;
+
     /**
-     * @param PaginatorFactoryInterface $paginatorFactory
-     * @param ProductsUpdater           $productsUpdater
-     * @param ProductsValidator         $productsValidator
-     * @param ProductsSaver             $productsSaver
-     * @param EventDispatcherInterface  $eventDispatcher
-     * @param ObjectDetacherInterface   $objectDetacher
+     * TODO @merge: on master remove PaginatorFactoryInterface from the constructor
+     * TODO @merge: on master remove ObjectDetacherInterface from the constructor
+     * TODO @merge: on master remove "= null" on ObjectManager $documentManager from the constructor
+     *
+     * @param PaginatorFactoryInterface  $paginatorFactory
+     * @param ProductsUpdater            $productsUpdater
+     * @param ProductsValidator          $productsValidator
+     * @param ProductsSaver              $productsSaver
+     * @param EventDispatcherInterface   $eventDispatcher
+     * @param ObjectDetacherInterface    $objectDetacher
+     * @param CacheClearerInterface|null $cacheClearer
+     * @param int                        $pageSize
      */
     public function __construct(
         PaginatorFactoryInterface $paginatorFactory,
@@ -62,7 +75,9 @@ class ProductRuleApplier implements ApplierInterface
         ProductsValidator $productsValidator,
         ProductsSaver $productsSaver,
         EventDispatcherInterface $eventDispatcher,
-        ObjectDetacherInterface $objectDetacher
+        ObjectDetacherInterface $objectDetacher,
+        CacheClearerInterface $cacheClearer = null,
+        $pageSize = 100
     ) {
         $this->paginatorFactory = $paginatorFactory;
         $this->productsUpdater = $productsUpdater;
@@ -70,6 +85,8 @@ class ProductRuleApplier implements ApplierInterface
         $this->productsSaver = $productsSaver;
         $this->eventDispatcher = $eventDispatcher;
         $this->objectDetacher = $objectDetacher;
+        $this->cacheClearer = $cacheClearer;
+        $this->pageSize = $pageSize;
     }
 
     /**
@@ -79,25 +96,47 @@ class ProductRuleApplier implements ApplierInterface
     {
         $this->eventDispatcher->dispatch(RuleEvents::PRE_APPLY, new SelectedRuleEvent($rule, $subjectSet));
 
-        $paginator = $this->paginatorFactory->createPaginator($subjectSet->getSubjectsCursor());
+        $productsPage = [];
+        foreach ($subjectSet->getSubjectsCursor() as $product) {
+            $productsPage[] = $product;
+            if (count($productsPage) >= $this->pageSize) {
+                $this->updateProducts($rule, $productsPage);
+                $productsPage = [];
+            }
+        }
 
-        foreach ($paginator as $productsPage) {
-            $this->productsUpdater->update($rule, $productsPage);
-            $validProducts = $this->productsValidator->validate($rule, $productsPage);
-            $this->productsSaver->save($rule, $validProducts);
-            $this->detachProducts($productsPage);
+        if (count($productsPage) > 0) {
+            $this->updateProducts($rule, $productsPage);
         }
 
         $this->eventDispatcher->dispatch(RuleEvents::POST_APPLY, new SelectedRuleEvent($rule, $subjectSet));
     }
 
     /**
+     * TODO @merge: on master rename and refactor method to only clear cache
+     *
      * @param array $productsPage
      */
-    protected function detachProducts(array $productsPage)
+    protected function detachProducts(array $productsPage = [])
     {
-        foreach ($productsPage as $product) {
-            $this->objectDetacher->detach($product);
+        if (null === $this->cacheClearer) {
+            foreach ($productsPage as $product) {
+                $this->objectDetacher->detach($product);
+            }
+        } else {
+            $this->cacheClearer->clear();
         }
+    }
+
+    /**
+     * @param RuleInterface $rule
+     * @param array         $products
+     */
+    protected function updateProducts(RuleInterface $rule, array $products)
+    {
+        $this->productsUpdater->update($rule, $products);
+        $validProducts = $this->productsValidator->validate($rule, $products);
+        $this->productsSaver->save($rule, $validProducts);
+        $this->detachProducts();
     }
 }
